requirements.txt
django
paho-mqtt
django-simpleui
channels
